add BloodPolicy

+++ app/Http/Controllers/Admin/Blood/BloodController.php
@@ -16,20 +16,27 @@ class BloodController extends Controller
 
     public function index(BloodFilter $filters)
     {
-        $bloods = Blood::latest()->filter($filters)->safePaginate();
+        $this->authorize('list', Blood::class);
+
+        $bloods = Blood::filter($filters)->latest()->safePaginate();
 
         return view('admin.blood.index', compact('bloods'));
     }
 
     public function create()
     {
+        $this->authorize('create', Blood::class);
+
         return view('admin.blood.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Blood::class);
+
         $request['mobile'] = make_mobile($request->mobile);
         $this->validateBlood($request);
+
         $blood = Blood::create($request->only(['mobile', 'city', 'blood_type', 'rh']));
 
         session_success(__('messages.blood.create', ['mobile' => $blood->mobile]));
@@ -39,13 +46,18 @@ class BloodController extends Controller
 
     public function edit(Blood $blood)
     {
+        $this->authorize('update', $blood);
+
         return view('admin.blood.edit', compact('blood'));
     }
 
     public function update(Request $request, Blood $blood)
     {
+        $this->authorize('update', $blood);
+
         $request['mobile'] = make_mobile($request->mobile);
         $this->validateBlood($request, true);
+
         $blood->update($request->only(['mobile', 'city', 'blood_type', 'rh']));
 
         session_success(__('messages.blood.update', ['mobile' => $blood->mobile]));
@@ -55,6 +67,8 @@ class BloodController extends Controller
 
     public function destroy(Blood $blood)
     {
+        $this->authorize('delete', $blood);
+
         $blood->delete();
 
         return api_success();
